format the arguments fmt cannot take as they are
fmt is type safe where the vararg StrFmt and DbgTrc were not, so five
arguments that a printf silently reinterpreted are now compile errors naming
type_is_unformattable_for.

Two enums under %d, EEntrenchSegmType and SObjectSelection::ESelectionType,
get static_cast<int>, the same treatment the trace call in FolderController.cpp
got earlier on this branch.

Three CDBIDs under a {} get ToString(). A CDBID is a struct; the vararg call
was pushing it whole and printf was reading a pointer out of the front of it,
so those two log lines never printed the id they claim to. The neighbouring
call in SimpleObjectState.cpp had ToString() all along.

A CPtr<IManipulator> under 0x%X was worse: printf read the smart pointer's
first word, which happens to be the pointer, and the %X truncated it to 32
bits. It is %p now, over the raw pointer cast to const void *, which is what
fmt's %p accepts and what the line meant to print.

+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapInfoEditorData_Selection.cpp
@@ -133,7 +133,7 @@ namespace NMapInfoEditor
 								bBackupCreated ? "true" : "false", vBackupPosition.x, vBackupPosition.y, vBackupPosition.z, fBackupDirection );
 		//
 		DebugTrace( "selection, type: %d, diffPos: ( %g, %g, %g ), diffDir: %g",
-								eSelectionType, vPositionDifference.x, vPositionDifference.y, vPositionDifference.z, fDirectionDifference );
+								static_cast<int>( eSelectionType ), vPositionDifference.x, vPositionDifference.y, vPositionDifference.z, fDirectionDifference );
 	}
 };
